+++ src/query/union/simple_union.rs
@@ -91,12 +91,12 @@ impl<TDocSet: DocSet> DocSet for SimpleUnion<TDocSet> {
         self.doc
     }
 
-    fn size_hint(&self) -> u32 {
+    fn size_hint(&self) -> u64 {
         self.docsets
             .iter()
             .map(|docset| docset.size_hint())
             .max()
-            .unwrap_or(0u32)
+            .unwrap_or(0)
     }
 
     fn count_including_deleted(&mut self) -> u32 {
